Day 2 of 100 day challange
The Thing-A-Ma-Jig is an exercise from the Head First Javascript book and I'm not really sure what it does but, I guess that was the point of the exercise. To try and figure out what's going on with the code. Here is the books description:

The Thing-A-Ma-Jig is quite a contraption-it clanks and clunks and even thunks, but what it really does, well, you've got us stumped. Coders claim they know how it works. Can you uncrack the code and find its quirks?